+++ lib/puma/const.rb
@@ -228,7 +228,6 @@ module Puma
     COLON = ": ".freeze
 
     NEWLINE = "\n".freeze
-    HTTP_INJECTION_REGEX = /[\r\n]/.freeze
 
     HIJACK_P = "rack.hijack?".freeze
     HIJACK = "rack.hijack".freeze
@@ -239,5 +238,13 @@ module Puma
     # Mininum interval to checks worker health
     WORKER_CHECK_INTERVAL = 5
 
+    # Illegal character in the key or value of response header
+    DQUOTE = "\"".freeze
+    HTTP_HEADER_DELIMITER = Regexp.escape("(),/:;<=>?@[]{}").freeze
+    ILLEGAL_HEADER_KEY_REGEX = /[\u0000-\u0025|#{DQUOTE}|#{HTTP_HEADER_DELIMITER}]/.freeze
+    ILLEGAL_HEADER_VALUE_REGEX = /[\000-\037]/.freeze
+
+    # Banned keys of response header
+    BANNED_HEADER_KEY = /rack.|status/.freeze
   end
 end